+++ src/main/java/com/geosegbar/infra/instrument/dtos/DeterministicLimitDTO.java
@@ -1,5 +1,7 @@
 package com.geosegbar.infra.instrument.dtos;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class DeterministicLimitDTO {
 
     private Long id;
-    private Double attentionValue;
-    private Double alertValue;
-    private Double emergencyValue;
+    private BigDecimal attentionValue;
+    private BigDecimal alertValue;
+    private BigDecimal emergencyValue;
 }
